[FIX] payment: fix access error in tx processing
The `callback_hash` and `callback_is_done` fields are not accessible
for non-admin users, so it will cause an access error whenever trying
to verify the hash inside `_execute_callback()`, and when marking the
callback execution successful.

Accessing it as `sudo` like other restricted transaction fields solves
the problem, and is safe, as the value is never leaked outside the
function.

Introduced in 5639960a7eb88f1f7b1aae0754905ad9fd9c8556.

closes odoo/odoo#77723

X-original-commit: a06dbaa8a76ffae27116a016bbfafcb45c852943
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>