[FIX][IMP] test_mail: be sure performance test include mail gateway
Purpose of query counter tests is to try to match real life use cases. In
mail those generally involve a correctly configured mail gateway. This is
why we set those parameters in performance tests, leading to a small increase
in some counters.

This is done in stable to better see evolution of those counters in various
stables.

Task-2661036
Prepares Task-36879 (MultiCompany Aliases)

X-original-commit: odoo/odoo@145aec0cea2080f6cb1b89e50f2a41dc1c116865
Part-of: odoo/odoo#77899