[FIX] mrp: control archive operations within form view
PR odoo/odoo#76376 made it so operations (of a BoM) can now be
(un)archived in the list view. This commit makes it so we can also do
the same in the form view as well as see the "Archived" ribbon in the
form view.

Part 3 of Task: 2660820

closes odoo/odoo#77772

Related: odoo/enterprise#21422
Signed-off-by: William Henrotin <Whenrow@users.noreply.github.com>