[FIX] website: allow to disable cache page based on content
In some case, we know that page will contains dynamic content and that cache
should be ignored.

E.g. as soon as a s_snippet_form will be present in page, we know that we need
the real csrf, and so we cannot use the cached response from another user.

Now, we disable the cache (cache_time=0)

closes odoo/odoo#77933

X-original-commit: d8e317f81ec886732f8ea19c3fcac32e0216dc1e
Signed-off-by: Jérémy Kersten (jke) <jke@openerp.com>