[FIX] crm: fix the rule based assignment
Bug
===
Since 324ffefd401d7ea3af0214dc6e588a215ff96a9c , if you have a team
with set to `assignment_max` and you try to automatically assign the
leads an error is raised.

Technical: in `_allocate_leads` we skip a team if `assignment_max` is
Falsy. So when we prepare the values for the notifications, the key
might not be present in the dict an error is raised.

Task-2658696

closes odoo/odoo#78043

X-original-commit: 1b600558ee095873e5b202513db58cabc24cbbd5
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>