[FIX] web_editor: restore overlay padding controls
Sometimes the padding controls disappeared for no apparent reason. This
was when going from a child element to a parent element -> following
BVR's investigation, it apparently appears that the UI is not updated
anymore since already visible and nothing should have to be updated
since [1].

While waiting for a deeper investigation about the onFocus / ui update
flows, this commit restores the padding controls another way: by simply
not hiding them and showing them again at each blur / focus. It does not
seem needed indeed... or at least I cannot find a reason why right now.
Indeed, it was not the case before [2] which changed the system without
any explanation. With [3], the onFocus part was later moved to updateUI
since each onFocus calls are now followed by an UI update. But in this
case, like before [2], it only seems needed to show or not the padding
controls at option initialization.

Deeper investigation and refactoring will follow.

[1]: https://github.com/odoo/odoo/commit/e2c5e3063f4c6091b2c65f2a0abcfb4b5ee052cf
[2]: https://github.com/odoo/odoo/commit/66cb7a4f1cc4febdd2d97107c5f5b1d9d97d71f7
[3]: https://github.com/odoo/odoo/commit/5810853da450f2d2de6fff91d6e6f2179669eb14

closes odoo/odoo#77911

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>