[FIX] core: one2many linking non-existing lines
The processing of the LINK command in a one2many should fail when the
line being linked does not exist.  However, there is one case where it
should not fail: when that line existed and was linked before applying
the commands.  That use-case may seem strange, but it actually exists:
deleting a line in a sales order automatically deletes the corresponding
reward lines.

closes odoo/odoo#78309

X-original-commit: 6c9e0afea8de4b2ca174ab55c69692bf9e09a03c
Signed-off-by: Raphael Collet (rco) <rco@openerp.com>