[FIX] auth_signup: bulk signup_url generation
When calling `_get_signup_url_for_action` on more than one partner,
because the `url` input parameter was re-used for the actual signup
URL before assigning to the partners map (in
8a1635b58628cba594a64b5c586ea7d860a785bc), from the second iteration
onwards the signup URL will almost certainly get misgenerated to
redirect to the previous signup URL (accumulating).

As `signup_url` is not normally accessed in bulk this should not
usually be an issue.

Reported by Andreas Brückl

closes odoo/odoo#78212

X-original-commit: d1457f158d10309eac30241f31f5edfc3f07578a
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>