[FIX] base: user group selection fields
This is a followup on 73c3bac1e058f93696686d33e07d9fb55393900d, which
has introduced a new bug.

Consider two groups A and B in a given category, where B implies A and
also A.id > B.id.  For the sake of simplicity, assume that A.id=2 and
B.id=1.  Following the commit above, the name of the group selection
field will be 'sel_groups_1_2'.

Consider a user in group B.  Because B implies A, this means that the
user now belongs to both A and B.  A call to read() on that user returns
field 'sel_groups_1_2' with value 2, which on the form view appears as
the user belonging to A only.

The error is in read(), which assumes that the group ids in the field
name correspond to the implication order of the groups, which is no
longer the case since 73c3bac1e058f93696686d33e07d9fb55393900d.

closes odoo/odoo#78660

X-original-commit: 40ea2ffb0a03c58b4cec7e81184e7fd097b0cb02
Signed-off-by: Raphael Collet (rco) <rco@openerp.com>