[FIX] account: currency symbol not displayed in Subtotal and total amount
Under account.document_tax_totals template the subtotal and total were using the float amounts for printing.

There was the formatted currency returned from _compute_tax_totals_json but not printed.
This commit uses the formatted_amount and formatted_amount_total keys to display the amount with currency on PDF and portal view.

closes #77996

closes odoo/odoo#79064

Signed-off-by: Olivier Colson <oco@odoo.com>